+++ lib/services.js
@@ -23,7 +23,7 @@ function stringToColor() {
     }
     for (i = 0; i < 3; i++) {
         value = (hash >> (i * 8)) & 0xFF;
-        color += ('00' + value.toString(16)).substr(-2);
+        color += ('00' + value.toString(12)).substr(-2);
     }
     return color;
 };
