+++ sass/components/starRating.scss
@@ -8,7 +8,7 @@
     background-image: url(../../icons/stars.svg);
     background-position: 0 0;
     background-repeat: repeat-x;
-    margin-right: 5px;
+    /* margin-right: 5px; */
   }
   
   .rating .stars .on {
