+++ sass/components/statistics.scss
@@ -2,6 +2,8 @@
     display: block;
     height: 100%;
     width: 100%;
+    padding: 8px;
+    overflow: hidden;
 }
 .statistic__header {
     display: flex;
